Condense cards into one folder and one stylesheet
