--- src/components/assets/cards/review-card/ReviewCard.css
@@ -1,100 +1,0 @@
-.card {
-    border: 1px solid #000;
-    min-width: 200px;
-    max-width: 300px;
-    min-height: 250px;
-    max-height: 350px;
-    overflow: hidden;
-    background-color: rgb(66, 76, 71);
-    box-shadow: 0 2px 5px black;
-}
-
-.card-content {
-    padding: 0.5rem;
-}
-
-.card-media-name {
-    padding: 0%;
-    background: none;
-    border: none;
-    text-align: left;
-    margin: 0;
-    font-size: 1.3rem;
-}
-
-.card-media-name:hover {
-    color: rgb(164, 255, 168);
-}
-
-.card-username {
-    padding: 0%;
-    background: none;
-    border: none;
-    text-align: left;
-    margin: 0.1rem 0;
-    font-size: 0.9rem;
-    color: grey;
-}
-
-.card-username-button {
-    padding: 0%;
-    background: none;
-    border: none;
-    text-align: left;
-    margin: 0.1rem 0;
-    font-size: 0.9rem;
-    color: grey;
-}
-
-.card-username-button:hover {
-    color: rgb(164, 255, 168);
-}
-
-.card-rating {
-    margin: 0.1rem 0;
-    font-size: 1.1rem;
-}
-
-.card-description {
-    font-size: 0.9rem;
-    color: #d0d0d0;
-    max-height: 6rem;
-    min-height: 8rem;
-    overflow-y: scroll;
-}
-
-/* Card Resizing for different screen sizes */
-
-.card:not(:last-child) {
-    margin-right: 4px;
-}
-
-@media (min-width: 500px) {
-    .card {
-        flex-basis: calc(50% - 10px);
-    }
-
-    .card:not(:last-child) {
-        margin-right: 4px;
-    }
-}
-
-@media (min-width: 700px) {
-    .card {
-        flex-basis: calc(calc(100% / 3) - 20px);
-    }
-
-    .card:not(:last-child) {
-        margin-right: 8px;
-    }
-}
-
-@media (min-width: 1100px) {
-    .card {
-        flex-basis: calc(25% - 30px);
-    }
-
-    .card:not(:last-child) {
-        margin-right: 12px;
-    }
-}
