+++ src/pages/search/search.css
@@ -12,7 +12,9 @@
   font-size: 0.9rem;
   margin: 0.5rem 0;
 }
-.status.error { color: #c00; }
+.status.error { 
+  color: #c00; 
+}
 
 .card-row {
   display: flex;
@@ -26,12 +28,14 @@
   height: 8px;
 }
 
-.media-row {}
-
 .user-row .card {
   width: 120px;
 }
+
 .user-row .card-image {
   height: 80px;
 }
-.user-row .card-title { font-size: 0.95rem; }
+
+.user-row .card-title { 
+  font-size: 0.95rem; 
+}
